Extract `PointerHook` from `FrameworkHook`
